# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LinkFlow API
  description: Advanced Microservices Workflow Automation Platform API
  version: 1.0.0
  contact:
    name: LinkFlow Team
    email: team@linkflow.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://api.linkflow.io/v1
    description: Production

security:
  - BearerAuth: []
  - ApiKeyAuth: []

tags:
  - name: Auth
    description: Authentication operations
  - name: Users
    description: User management
  - name: Workflows
    description: Workflow operations
  - name: Executions
    description: Workflow execution operations
  - name: Nodes
    description: Node management
  - name: Credentials
    description: Credential management
  - name: Schedules
    description: Schedule management
  - name: Analytics
    description: Analytics and metrics

paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /workflows:
    get:
      tags: [Workflows]
      summary: List workflows
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, error]
      responses:
        '200':
          description: List of workflows
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workflow'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
    post:
      tags: [Workflows]
      summary: Create workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '201':
          description: Workflow created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'

  /workflows/{id}:
    get:
      tags: [Workflows]
      summary: Get workflow by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Workflows]
      summary: Update workflow
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkflowRequest'
      responses:
        '200':
          description: Workflow updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
    delete:
      tags: [Workflows]
      summary: Delete workflow
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Workflow deleted

  /workflows/{id}/execute:
    post:
      tags: [Workflows]
      summary: Execute workflow
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                data:
                  type: object
                  description: Input data for execution
      responses:
        '202':
          description: Execution started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'

  /executions:
    get:
      tags: [Executions]
      summary: List executions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: workflowId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed]
      responses:
        '200':
          description: List of executions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Execution'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

  /executions/{id}:
    get:
      tags: [Executions]
      summary: Get execution by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Execution details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'

  /executions/{id}/stop:
    post:
      tags: [Executions]
      summary: Stop execution
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Execution stopped
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Execution'

  /nodes/types:
    get:
      tags: [Nodes]
      summary: List node types
      responses:
        '200':
          description: List of node types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeType'

  /analytics/dashboard:
    get:
      tags: [Analytics]
      summary: Get analytics dashboard
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        status:
          type: string
          enum: [active, inactive, suspended]
        createdAt:
          type: string
          format: date-time

    Workflow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/Connection'
        status:
          type: string
          enum: [active, inactive, error]
        isActive:
          type: boolean
        version:
          type: integer
        createdAt:
          type: string
          format: date-time

    Node:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        position:
          type: object
          properties:
            x:
              type: number
            y:
              type: number
        parameters:
          type: object

    Connection:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string

    Execution:
      type: object
      properties:
        id:
          type: string
          format: uuid
        workflowId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, running, completed, failed]
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        executionTime:
          type: integer
        data:
          type: object
        error:
          type: string

    NodeType:
      type: object
      properties:
        type:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        schema:
          type: object

    Dashboard:
      type: object
      properties:
        totalWorkflows:
          type: integer
        totalExecutions:
          type: integer
        activeWorkflows:
          type: integer
        runningExecutions:
          type: integer
        successRate:
          type: number

    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    CreateWorkflowRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/Connection'

    UpdateWorkflowRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/Node'
        connections:
          type: array
          items:
            $ref: '#/components/schemas/Connection'

    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'